requirements.txt
Flask
Flask-SocketIO
gunicorn
eventlet
redis
python-dotenv